`main` 周辺の責務を分割し、自動テスト実行導線を整理
## 概要

`main.ts` に集まっていた純ロジックと描画補助を分割し、自動テストを用途別のファイルへ再編しました。あわせて、テスト実行導線を `scripts/run-tests.mjs` ベースへ整理しています。

## 変更内容

### `main` 周辺の分割

- `src/ts/ai-json-util.ts` を追加
- `src/ts/main-render.ts` を追加
- `src/ts/main-util.ts` を追加
- `src/ts/main.ts` から上記モジュールへ処理を切り出し
- 生成物として `src/js/ai-json-util.js`、`src/js/main-render.js`、`src/js/main-util.js` を追加
- `scripts/build-project.mjs` と `mikuproject-src.html` を更新し、追加モジュールをビルド・読込対象に反映

### 自動テストの再編

- `tests/mikuproject-main.test.js` を大幅に縮小
- 以下の用途別テストファイルを追加
  - `tests/mikuproject-ai-json-util.test.js`
  - `tests/mikuproject-main-ai-json.test.js`
  - `tests/mikuproject-main-file-input-wiring.test.js`
  - `tests/mikuproject-main-preview-export.test.js`
  - `tests/mikuproject-main-util.test.js`
  - `tests/mikuproject-main-validation.test.js`
  - `tests/mikuproject-main-xlsx-import.test.js`
  - `tests/mikuproject-msproject-xml-roundtrip.test.js`

### テスト実行導線の整理

- `scripts/run-tests.mjs` を更新
- `package.json` の test script を整理
  - `test`
  - `test:fast`
  - `test:ui`
  - `test:full`

### ドキュメント更新

- `docs/TODO.md` から完了済みの自動テスト改善項目を削除

## 更新ファイル

- `docs/TODO.md`
- `mikuproject-src.html`
- `mikuproject.html`
- `package.json`
- `scripts/build-project.mjs`
- `scripts/run-tests.mjs`
- `src/ts/main.ts`
- `src/ts/ai-json-util.ts`
- `src/ts/main-render.ts`
- `src/ts/main-util.ts`
- `src/ts/msproject-xml.ts`
- `src/js/main.js`
- `src/js/ai-json-util.js`
- `src/js/main-render.js`
- `src/js/main-util.js`
- `src/js/msproject-xml.js`
- `tests/mikuproject-main.test.js`
- 追加された各テストファイル
