+++ mikuproject-src.html
@@ -434,6 +434,8 @@
   </main>
   <script src="src/js/types.js"></script>
   <script src="src/js/markdown-escape.js"></script>
+  <script src="src/js/ai-json-util.js"></script>
+  <script src="src/js/main-util.js"></script>
   <script src="src/js/excel-io.js"></script>
   <script src="src/js/msproject-xml.js"></script>
   <script src="src/js/project-workbook-schema.js"></script>
@@ -443,6 +445,7 @@
   <script src="src/js/wbs-xlsx.js"></script>
   <script src="src/js/wbs-markdown.js"></script>
   <script src="src/js/wbs-svg.js"></script>
+  <script src="src/js/main-render.js"></script>
   <script src="src/js/main.js"></script>
 </body>
 </html>
